AddTask now supports functions with parameters
Added AddTask overloads for functions with any number and types of parameters (thanks to parameter packs), as well as for lambdas (via std::move).
Changed AsyncTask to work with std::function&& instead of a link to function to prevent crits when passing functions with parameters.

Known issues: currently wait_all(); method seems to be bugged and is not working properly.